Update Readme.md
The First Paper aims to unlock the Potential of event cameras when combined with a deep neural network approach for a challenging motion-estimation task (prediction of a vehicle's steering angle).

The Second paper introduces a framework for LiDAR localization that utilizes machine learning and is specifically tailored for self-driving car applications called L3-Net. This new LiDAR localization system employs machine learning and achieves accuracy in centimeter-level localization, which is on par with previous cutting-edge systems that used manual pipelines.